Name testing and new test cases
I changed the name of the tests we wrote during the lab, I also added
some new test, there's a total of 17 test of which 6 failed. Most of the
failures are due to overflow, However there's
centralValueCloseToMaxRangeTest2 which uses values under the the max
double values and it still fails.
When picking the functions to test, make sure to read the documentaion
in order to avoid creating tests due to misunderstandings like the prof
mentioned today in class. (We thought of testing range by giving it
reverse values new Range(1,-1) while the documentation specified that
the first argument was the lower bound and the second was the upper
bound where lower must be <= upper and upper must be >= lower)